Move processing type determining logic to its own registry
